Added extension
and create folders for support
